avito api/requirements.txt
requests==2.32.5
pytest==9.0.2
allure-pytest==2.15.3
pydantic==2.12.5
ruff==0.15.8
black==26.3.1
isort==8.0.1
